chore(chips): Moved delete tests from chip-area spec to chip spec.
